improve the fmt::Debug implementations in public types
It switches the `Debug` implementations in some public types to the manual,
instead of using `#[derive(Debug)]`.
This is also intended to hide the implementation details around the
message data from the kernel and to prevent unexpected leakage of the
unprocessed data to the user side, and hence it has more than just to
improve the readability of debugging log messages.

At the same time, the unused `Debug` derivations and the tracing output
inside the library are removed.