polish operation traits
The contexts provided by the runtime are all aggregated in each
operation trait.

Additionally, the traits `Filesystem`/`LocalFilesystem` that represent
sets of operation callbacks are deleted. I am currently considering the
implementation pattern for writing filesystem that do not use these
traits.

polyfuse-smol is also temporarily disabled as it relies on the old
filesystem abstraction.  Rewriting the new filesystem daemon
implementation will be needed as soon as the new filesystem
representation format is fixed.